New security feature: Change Set Name Prefix
ChangeSets have names, and the ARN for a change set has nothing to do with the stack name. Thus if you want to lock down a role to only be able to execute change sets for which it created or has access to, you must do it off of a naming convention. This pattern allows a change set name to be prefixed by a given name.