Fix: Ability to delete change set that didn't apply
If a change set is not applied, the automation cannot run again with a changeset by the same name unless the previous change set is first deleted. This occurs more frequently than you might expect, as when somebody submits a change that ultimately no change was needed - you end up with a change set in a fail state since there is no change to apply. The next time this stack is updated, the automation first must delete that change set if it exists before trying to create the new change set to execute.