Fix bug with deleting stacks where template is gone too
Bumped into a bug where if the template file was deleted prior to the stack file being deleted, the stack would fail to delete. We only need to test for the template file being available on create / update operations. On delete, we don't need the template file - just the stack name so we can delete it.