Final set of changes for Multi-region support:
CloudFormation when publishing to an SNS topic attached to the stack, can only publish to a topic that is in the same region as the stack itself. To overcome this, you must create an SNS topic in each region you want the deployer to be able to deploy to using the `multi-region-template.yaml` stack.  The main stack that creates GitFormation itself, will setup Lambda permissions to allow these SNS topics to subscribe the Slack CF event lambda directly to the topics in the regions.

As new regions come on-line, this will require some manual touch of the GitFormation SAM stack file due to a limitation in Lambda Permission policies not allowing a wildcard ARN pattern match.