Add support for multi-region deployments (#21)
* Add support for multi-region deployments

* Changes the stack path naming to now require that the folder under the account is now a region. The folder name must parse into a valid region name for the launch to succeed.
* Updated the semanticStackName parser for the stack name to now look at the forth group since we added the region name prior to the stack path itself now.
* Added new simple function to parse out the region id
* Moved the cfClientBuilder to a separate function so we can write a test to verify that when a region is passed, that it is used in building the client.
* Updated tests to support updates.

* Update the references to the name of the project to GitFormation

* Fix bug with deleting stacks where template is gone too

Bumped into a bug where if the template file was deleted prior to the stack file being deleted, the stack would fail to delete. We only need to test for the template file being available on create / update operations. On delete, we don't need the template file - just the stack name so we can delete it.

* Final set of changes for Multi-region support:

CloudFormation when publishing to an SNS topic attached to the stack, can only publish to a topic that is in the same region as the stack itself. To overcome this, you must create an SNS topic in each region you want the deployer to be able to deploy to using the `multi-region-template.yaml` stack.  The main stack that creates GitFormation itself, will setup Lambda permissions to allow these SNS topics to subscribe the Slack CF event lambda directly to the topics in the regions.

As new regions come on-line, this will require some manual touch of the GitFormation SAM stack file due to a limitation in Lambda Permission policies not allowing a wildcard ARN pattern match.