Feature: Support kicking off from GitHub webhook
This enables CodePipeline to start from GitHub Webhooks instead of pulling GitHub every minute or two. This will speed up the time between a PR merge on an infrastructure change and seeing the change(s) start to happen.

The value you need to provide is a secret that YOU generate. AWS informs GitHub to use that value on a webhook request to CodePipeline.